modularized components in server
